Refactor ioutil
