+++ internal/cmdutil/exec.go
@@ -1,4 +1,7 @@
-package ioutil
+// Copyright 2023 Jetpack Technologies Inc and contributors. All rights reserved.
+// Use of this source code is governed by the license in the LICENSE file.
+
+package cmdutil
 
 import (
 	"bytes"
